Fix strikethrough positioning to match Zed and Neovide using proper font metrics
- Use font's built-in strikeout_offset when available (primary method)
- Fall back to x_height/2 above baseline (typographically correct)
- Final fallback to approximate position at 30% of line height
- Add x_height to font metrics pipeline for accurate positioning
- Reduce default strikethrough thickness from 2.0 to 1.5 for better appearance
- Add comprehensive tests for all positioning scenarios
- Extract underline creation logic into testable helper method

This matches how modern terminals like Zed and Neovide position strikethrough
text decorations using proper typographic principles.

Fix cursor block size to be based on font metrics instead of line height

- Add ascent to font metrics pipeline (TextRunStyle, RunData, render_data)
- Update cursor rendering to use font_height (ascent + descent) instead of line_height
- Calculate cursor_top as baseline - ascent for proper positioning
- Cursor size now remains consistent regardless of line height increases
- Add comprehensive tests for cursor font-based sizing
- Test different font metrics combinations and line heights
- Verify cursor positioning with various ascent/descent ratios

This ensures that increasing line height for better readability doesn't make
the cursor unnecessarily tall, matching the behavior of modern text editors.

Clean up tests and comments

- Remove println! statements from tests for cleaner output
- Remove references to other editors from comments
- Update cursor height calculations in tests to use font_height instead of line_height
- Maintain focus on Rio's own implementation and standards

+++ sugarloaf/src/components/rich_text/batch.rs
@@ -17,7 +17,7 @@ use bytemuck::{Pod, Zeroable};
 #[derive(Default, Clone, Copy)]
 pub struct RunUnderline {
     pub enabled: bool,
-    pub offset: i32,
+    pub offset: f32,
     pub size: f32,
     pub color: [f32; 4],
     pub is_doubled: bool,
@@ -3805,7 +3805,7 @@ impl BatchManager {
             let ux = x;
             // Position underline below baseline by adding the calculated offset
             // This ensures proper underline placement in the descent area
-            let uy = baseline + underline.offset as f32;
+            let uy = baseline + underline.offset;
 
             let end = x + advance;
             if ux < end {
